requirements.txt
flask
flask-cors
gspread
google-auth
pandas
numpy
fpdf2
pytz
python-dotenv
pyjwt
cryptography
requests
openpyxl

# Simulador Streamlit (diresimulator.py / simulador_dv)
streamlit>=1.28
st-gsheets-connection
streamlit-folium
altair
folium
Pillow

# API web (HTML/CSS/JS + FastAPI)
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2
httpx>=0.26